#8: Parsed method invocations, expression statements, object createion expressions.

+++ src/io/jsrminer/uml/mapping/LeafStatementMapping.java
@@ -84,18 +84,18 @@ public class LeafStatementMapping extends StatementMapping implements Comparable
         } else if (parent1 == null && parent2 != null) {
             String s2 = parent2.getText();
             int distance = StringDistance.editDistance("{", s2);
-            double normalized = (double) distance / (double) Math.max(1, s2.length());
+            double normalized = (double) distance / Math.max(1, s2.length());
             return normalized;
         } else if (parent1 != null && parent2 == null) {
             String s1 = parent1.getText();
             int distance = StringDistance.editDistance(s1, "{");
-            double normalized = (double) distance / (double) Math.max(s1.length(), 1);
+            double normalized = (double) distance / Math.max(s1.length(), 1);
             return normalized;
         }
         String s1 = parent1.getText();
         String s2 = parent2.getText();
         int distance = StringDistance.editDistance(s1, s2);
-        double normalized = (double) distance / (double) Math.max(s1.length(), s2.length());
+        double normalized = (double) distance / Math.max(s1.length(), s2.length());
         return normalized;
     }
 }
